+++ src/StorageKey/IdentityErrorStorageKey.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey;
+
+use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\IdentityErrorKeyInterface;
+
+class IdentityErrorStorageKey extends AbstractStorageKey implements IdentityErrorKeyInterface
+{
+}